requirements.txt
alembic==1.8.1
aniso8601==9.0.1
bcrypt==4.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==37.0.4
Flask==2.2.2
Flask-API==3.0.post1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.4
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-Pydantic==0.11.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
greenlet==1.1.3.post0
idna==3.3
importlib-metadata==5.0.0
importlib-resources==5.10.0
itsdangerous~=2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
Naked==0.1.31
packaging==21.3
psycopg2-binary==2.9.4
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.2
PyJWT==2.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.5
PyYAML==6.0
razorpay==1.3.0
requests==2.28.1
shellescape==3.8.1
six==1.16.0
SQLAlchemy==1.4.42
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.9.0

stripe~=4.2.0